pyproject.toml
[project]
name = "advanced-seleniumlibrary-workshop-robocon2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bottle==0.13.2",
    "robotframework==7.1.1",
    "robotframework-requests==0.9.7",
    "robotframework-seleniumlibrary==6.7.0",
]
